+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Below you can find the changes included in each release.
 
+## v4.4.2
+
+- Design tweaks
+- Catalogue listing totals
+
 ## v4.4.1
 
 - bugfix: automatic bookmarking doesn't fall over when no config is provided